Modify int to long in interface, routes and repo

+++ src/main/kotlin/com/example/dao/Customers.kt
@@ -5,8 +5,8 @@ import com.example.models.CustomerCreation
 
 interface Customers {
     suspend fun getAll(): List<CustomerCreated>
-    suspend fun get(id: Int): CustomerCreated?
+    suspend fun get(id: Long): CustomerCreated?
     suspend fun add(customerCreation: CustomerCreation): CustomerCreated?
     suspend fun edit(customerCreated: CustomerCreated): Boolean
-    suspend fun delete(id: Int): Boolean
+    suspend fun delete(id: Long): Boolean
 }
